closes #7, #8, #9, #10. Added LanguagePolicty. Added custom notifications. Added DIRECTORY_SEPARATOR

+++ src/app/Http/Controllers/LocalisationController.php
@@ -60,6 +60,8 @@ class LocalisationController extends Controller
 
     public function destroy(Language $localisation)
     {
+        $this->authorize('destroy', $localisation);
+
         return $this->localisation->destroy($localisation);
     }
 }
